+++ account-kit/smart-contracts/src/ma-v2/modules/webauthn-validation/abis/webauthnValidationAbi.ts
